Add files via upload
This is a project about a program to calculate a worker’s wage based on the number of meters he has built in length and width. This is done in two steps. The first step is multiplying the length by the width. the second step is multiplying the result by the amount that the worker receives per meter. Finally, the wage of the worker will be calculated perfectly.